Add files via upload
This contains 2 steps:
1. Create application image using below docker command:
  docker build -t my-go-app .
2. Then run the docker compose file to run the whole system using below command.
  docker-compose up -d
