+++ display.go
@@ -1,0 +1,34 @@
+package main
+
+import (
+	"context"
+	"encoding/json"
+	"fmt"
+	"net/http"
+
+	"github.com/gorilla/mux"
+	"github.com/lovoo/goka"
+)
+
+func runView(initialized chan struct{}) {
+	<-initialized
+
+	view, err := goka.NewView(brokers,
+		goka.GroupTable(group),
+		new(userCodec),
+	)
+	if err != nil {
+		panic(err)
+	}
+
+	root := mux.NewRouter()
+	root.HandleFunc("/features/{key}", func(w http.ResponseWriter, r *http.Request) {
+		value, _ := view.Get(mux.Vars(r)["key"])
+		data, _ := json.Marshal(value)
+		w.Write(data)
+	})
+	fmt.Println("View opened at http://localhost:9095/")
+	go http.ListenAndServe(":9095", root)
+
+	view.Run(context.Background())
+}
