+++ docker.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+docker-compose up -d
