+++ go.mod
@@ -3,19 +3,7 @@ module kafka-basics
 go 1.16
 
 require (
-	github.com/Shopify/sarama v1.33.0
-	github.com/Shopify/toxiproxy/v2 v2.4.0 // indirect
-	github.com/go-stack/stack v1.8.0
-	github.com/golang/mock v1.4.4
 	github.com/gorilla/mux v1.8.0
-	github.com/hashicorp/go-multierror v1.1.1
-	github.com/klauspost/compress v1.15.6 // indirect
-	github.com/lovoo/goka v1.1.6 // indirect
-	github.com/pierrec/lz4/v4 v4.1.14 // indirect
-	github.com/stretchr/testify v1.7.1
-	github.com/syndtr/goleveldb v1.0.0
-	golang.org/x/net v0.0.0-20220520000938-2e3eb7b945c2 // indirect
-	golang.org/x/sync v0.0.0-20201207232520-09787c993a3a
-	gopkg.in/redis.v5 v5.2.9
-	gopkg.in/yaml.v2 v2.4.0
+	github.com/linkedin/goavro v2.1.0+incompatible
+	github.com/lovoo/goka v1.1.6
 )
